social-links-profile-main-frontend-mentor
I created a social links profile, aiming to replicate the provided design as closely as possible. I used various tools and technologies at my disposal to complete the challenge. It was a great opportunity to practice and explore new skills that I wanted to improve. I felt free to experiment and apply my knowledge to ensure the design was as faithful as possible to the given example.